BIOINF595w25/labs/lab_06_docking_hit_picking/src/molecule_embedding_requirements.txt
absl-py==2.1.0
addict==2.4.0
adjustText==1.3.0
aimsim_core==2.2.2
aiobotocore==2.19.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aioitertools==0.12.0
aiosignal==1.3.1
anndata==0.11.3
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_api_compat==1.10.0
arrow==1.3.0
astartes==1.3.0
asttokens==3.0.0
async-lru==2.0.4
attrs==23.2.0
autocommand==2.2.2
babel==2.17.0
backports.entry-points-selectable==1.3.0
backports.tarfile==1.2.0
beautifulsoup4==4.13.3
biopython==1.84
bleach==6.2.0
blitzgsea==1.3.47
bokeh==3.6.3
boto3==1.35.53
botocore==1.36.3
cachetools==5.5.1
certifi==2024.7.4
cffi==1.17.1
charset-normalizer==3.3.2
-e git+https://github.com/chemprop/chemprop.git@fa2d6dceb054ced6fe46191a628cfd664037d5e7#egg=chemprop
chex==0.1.88
click==8.1.8
cloudpathlib==0.20.0
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.1
cycler==0.12.1
datamol==0.12.5
datasets==2.21.0
datashader==0.17.0
debugpy==1.8.12
decorator==5.1.1
decoupler==1.8.0
defusedxml==0.7.1
descriptastorus==2.8.0
dill==0.3.8
distlib==0.3.8
docker-pycreds==0.4.0
docrep==0.3.2
equinox==0.11.11
et_xmlfile==2.0.0
etils==1.12.0
evaluate==0.4.3
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.13.1
flax==0.10.3
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2025.2.0
future==1.0.0
gcsfs==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
google-api-core==2.24.1
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-cloud-core==2.4.1
google-cloud-storage==3.0.0
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.70.0
h11==0.14.0
h2o==3.46.0.6
h5py==3.12.1
holoviews==1.20.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.24.6
humanize==4.12.1
hypothesis==6.93.1
idna==3.8
igraph==0.11.8
importlib==1.0.4
importlib_metadata==8.5.0
importlib_resources==6.5.2
inflect==7.3.1
iniconfig==2.0.0
iopath==0.1.10
ipykernel==6.29.5
ipython==8.32.0
isoduration==20.11.0
jaraco.collections==5.1.0
jaraco.context==5.3.0
jaraco.functools==4.0.1
jaraco.text==3.12.1
jax==0.5.0
jaxlib==0.5.0
jaxopt==0.8.3
jaxtyping==0.2.38
jedi==0.19.2
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lamin_utils==0.13.10
legacy-api-wrap==1.4.1
leidenalg==0.10.2
lightning==2.4.0
lightning-utilities==0.11.9
lineax==0.0.7
linkify-it-py==2.0.3
llvmlite==0.43.0
lmdb==1.6.2
loguru==0.7.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mergedeep==1.3.4
mhfp==1.9.6
mistune==3.1.2
ml-dtypes==0.4.0
ml_collections==1.0.0
molfeat==0.10.1
molflux==0.8.0
MolVS==0.1.1
mordredcommunity==2.0.6
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
mudata==0.3.1
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.16
muon==0.1.7
natsort==8.4.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
numpyro==0.17.0
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
openpyxl==3.1.5
opt-einsum==3.3.0
optax==0.2.4
orbax-checkpoint==0.11.6
ott-jax==0.5.0
overrides==7.7.0
packaging==24.2
padelpy==0.1.16
pandas==2.2.3
pandas-flavor==0.6.0
pandocfilters==1.5.1
panel==1.6.1
param==2.2.0
parso==0.8.4
patsy==1.0.1
pertpy==0.9.5
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.2.2
pluggy==1.3.0
ply==3.11
pmapper==1.1.2
portalocker==3.1.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
proto-plus==1.26.0
protobuf==5.29.3
psutil==6.1.0
ptyprocess==0.7.0
PubChemPy==1.0.4
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyct==0.5.0
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pynndescent==0.5.13
Pyomo==6.8.2
pyparsing==3.2.1
pyro-api==0.1.2
pyro-ppl==1.9.1
pytest==7.4.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytorch-lightning==2.4.0
pytz==2024.1
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq==26.2.1
RapidFuzz==3.10.1
rdkit==2024.3.6
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
s3fs==2025.2.0
s3transfer==0.10.3
scanpy==1.11.0
scikit-learn==1.5.2
scikit-misc==0.5.1
scipy==1.14.1
scvi-tools==1.2.2.post2
seaborn==0.13.2
selfies==2.2.0
Send2Trash==1.8.3
sentry-sdk==2.22.0
session-info2==0.1.2
setproctitle==1.3.4
setuptools==75.6.0
simplejson==3.20.1
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
sparse==0.15.5
sparsecca==0.3.1
stack-data==0.6.3
statsmodels==0.14.4
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tensorstore==0.1.71
terminado==0.18.1
texttable==1.7.0
thefuzz==0.22.1
threadpoolctl==3.5.0
tinycss2==1.4.0
tmtools==0.2.0
tokenizers==0.21.0
tomli==2.0.1
toolz==1.0.0
torch==2.6.0
torchaudio==2.6.0
torchmetrics==1.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.66.5
traitlets==5.14.3
treescope==0.1.9
triton==3.2.0
typeguard==4.3.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
umap-learn==0.5.7
unicore==0.0.1
unimol_tools==0.1.2.post2
uri-template==1.3.0
urllib3==2.2.2
virtualenv==20.25.0
wadler_lindig==0.1.3
wandb==0.19.6
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.43.0
wrapt==1.17.2
xarray==2024.11.0
xgboost==2.1.4
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.9.4
zipp==3.20.2
